+++ routes/orderRoutes/orderRoutes.js
@@ -2,6 +2,10 @@ import express from 'express';
 import OrderModel from '../../models/orders.js';
 import { isUserAuthorized } from '../../utils/authUtils.js';
 
+import Stripe from 'stripe';
+
+const stripe = Stripe("your-secret-key");
+
 const router = express.Router();
 
 router.post('/order', isUserAuthorized, async (req, res) => {
@@ -37,6 +41,30 @@ router.post('/order', isUserAuthorized, async (req, res) => {
       success: false,
       error: 'Internal Server Error',
     });
+  }
+});
+
+router.post("/create-payment-intent", async (req, res) => {
+  try {
+    const { amount, currency } = req.body;
+
+    // Validate the input
+    if (!amount || !currency) {
+      return res.status(400).json({ error: "Amount and currency are required" });
+    }
+
+    // Create the PaymentIntent
+    const paymentIntent = await stripe.paymentIntents.create({
+      amount, // Amount in the smallest currency unit (e.g., cents for USD)
+      currency,
+      payment_method_types: ["card"],
+    });
+
+    // Respond with the client secret
+    res.status(200).json({ clientSecret: paymentIntent.client_secret });
+  } catch (error) {
+    console.error("Error creating PaymentIntent:", error);
+    res.status(500).json({ error: "Internal server error" });
   }
 });
